[refactor] Extract topological_sort to separate analysis module. [python/assassyn/analysis/topo.py] Create new module with topological_sort function for module dependency analysis. [python/assassyn/codegen/verilog/top.py] Import topological_sort from analysis.topo and remove local implementation. Remove deque import as it's no longer needed. This refactoring improves code organization by moving the topological sort algorithm to the analysis module where other dependency analysis utilities are located.
